+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/model/metamodel/gen/PersistentPropertyMatcher.java
@@ -1,3 +1,20 @@
+/*
+ * JOPA
+ * Copyright (C) 2024 Czech Technical University in Prague
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library.
+ */
 package cz.cvut.kbss.jopa.model.metamodel.gen;
 
 import cz.cvut.kbss.jopa.utils.EntityPropertiesUtils;
@@ -43,5 +60,24 @@ public abstract class PersistentPropertyMatcher<T extends MethodDescription> ext
             }
         } while (type != null);
         return Optional.empty();
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof PersistentPropertyMatcher<?> that)) {
+            return false;
+        }
+        if (!super.equals(o)) {
+            return false;
+        }
+        return Objects.equals(parentType, that.parentType);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(super.hashCode(), parentType);
     }
 }