+++ ontodriver-jena/src/main/java/cz/cvut/kbss/ontodriver/jena/list/ReferencedListHelper.java
@@ -1,3 +1,20 @@
+/*
+ * JOPA
+ * Copyright (C) 2024 Czech Technical University in Prague
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library.
+ */
 package cz.cvut.kbss.ontodriver.jena.list;
 
 import cz.cvut.kbss.ontodriver.jena.util.JenaUtils;
@@ -16,10 +33,9 @@ class ReferencedListHelper {
     }
 
     static Stream<RDFNode> toRdfNodes(Object value, Assertion nodeContentAssertion) {
-        if (value instanceof MultilingualString) {
-            final MultilingualString mls = (MultilingualString) value;
+        if (value instanceof MultilingualString mls) {
             return mls.getValue().entrySet().stream()
-                      .map((e) -> JenaUtils.valueToRdfNode(nodeContentAssertion, new Value<>(new LangString(e.getValue(), e.getKey()))));
+                      .map(e -> JenaUtils.valueToRdfNode(nodeContentAssertion, new Value<>(new LangString(e.getValue(), e.getKey()))));
         } else {
             return Stream.of(JenaUtils.valueToRdfNode(nodeContentAssertion, new Value<>(value)));
         }