+++ ontodriver-rdf4j/src/main/java/cz/cvut/kbss/ontodriver/rdf4j/list/SimpleListIterator.java
@@ -1,6 +1,6 @@
 /*
  * JOPA
- * Copyright (C) 2023 Czech Technical University in Prague
+ * Copyright (C) 2024 Czech Technical University in Prague
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -98,7 +98,7 @@ class SimpleListIterator extends AbstractListIterator<NamedResource> {
         return createAxiom(current.getSubject(), assertion, (Resource) current.getObject());
     }
 
-    private Axiom<NamedResource> createAxiom(Resource subject, Assertion assertion, Resource value) {
+    private static Axiom<NamedResource> createAxiom(Resource subject, Assertion assertion, Resource value) {
         return new AxiomImpl<>(NamedResource.create(subject.stringValue()), assertion, new Value<>(NamedResource.create(value.stringValue())));
     }